#141: Initial attempt at separating the gRPC stuff to a separate project
