+++ Yaref92.Events.sln
@@ -24,6 +24,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventMessenger", "samples\E
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yaref92.Events.Transport.Tcp", "src\Yaref92.Events.Transport.Tcp\Yaref92.Events.Transport.Tcp.csproj", "{18B2F85C-58D1-4F45-BB2A-8E7731AA46EC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yaref92.Events.Transport.Grpc", "src\Yaref92.Events.Transport.Grpc\Yaref92.Events.Transport.Grpc.csproj", "{5AB9FAD9-6BF3-479D-B215-2927833BB518}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -60,6 +62,10 @@ Global
 		{18B2F85C-58D1-4F45-BB2A-8E7731AA46EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{18B2F85C-58D1-4F45-BB2A-8E7731AA46EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{18B2F85C-58D1-4F45-BB2A-8E7731AA46EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5AB9FAD9-6BF3-479D-B215-2927833BB518}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5AB9FAD9-6BF3-479D-B215-2927833BB518}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5AB9FAD9-6BF3-479D-B215-2927833BB518}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5AB9FAD9-6BF3-479D-B215-2927833BB518}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
